+++ Saule/Http/JsonApiAttribute.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Net.Http;
 using System.Web.Http.Filters;
 
 namespace Saule.Http
@@ -7,7 +8,7 @@ namespace Saule.Http
     /// An optional attribute that can be used to opt an api into returning a JsonApi response.
     /// </summary>
     [AttributeUsage(AttributeTargets.Method | AttributeTargets.Class)]
-    public class JsonApiAttribute : ActionFilterAttribute
+    public sealed class JsonApiAttribute : ActionFilterAttribute
     {
         /// <summary>
         /// See base class documentation.
@@ -15,7 +16,23 @@ namespace Saule.Http
         /// <param name="context">The action context.</param>
         public override void OnActionExecuted(HttpActionExecutedContext context)
         {
-            JsonApiProcessor.ProcessRequest(context);
+            var config = new JsonApiConfiguration();
+            JsonApiProcessor.ProcessRequest(context.Request, context.Response, config, requiresMediaType: false);
+
+            if (context.Exception != null)
+            {
+                return;
+            }
+
+            var responseContent = context.Response.Content as ObjectContent;
+            if (responseContent == null)
+            {
+                return;
+            }
+
+            var formatter = new JsonApiMediaTypeFormatter(context.Request, config);
+
+            context.Response.Content = new ObjectContent(responseContent.ObjectType, responseContent.Value, formatter);
         }
     }
 }
